refactor: remove authorized email trait in favour of rule
